+++ core/lti/database.php
@@ -13,7 +13,7 @@ $DATABASE_UNINSTALL = array(
 $DATABASE_INSTALL = array(
 array( "{$CFG->dbprefix}lti_key",
 "create table {$CFG->dbprefix}lti_key (
-    key_id              MEDIUMINT NOT NULL AUTO_INCREMENT,
+    key_id              INTEGER NOT NULL AUTO_INCREMENT,
     key_sha256          CHAR(64) NOT NULL UNIQUE,
     key_key             VARCHAR(4096) NOT NULL,
 
@@ -29,11 +29,11 @@ array( "{$CFG->dbprefix}lti_key",
 
 array( "{$CFG->dbprefix}lti_context",
 "create table {$CFG->dbprefix}lti_context (
-    context_id          MEDIUMINT NOT NULL AUTO_INCREMENT,
+    context_id          INTEGER NOT NULL AUTO_INCREMENT,
     context_sha256      CHAR(64) NOT NULL,
     context_key         VARCHAR(4096) NOT NULL,
 
-    key_id              MEDIUMINT NOT NULL, 
+    key_id              INTEGER NOT NULL, 
 
     title               VARCHAR(2048) NULL,
 
@@ -52,11 +52,11 @@ array( "{$CFG->dbprefix}lti_context",
 
 array( "{$CFG->dbprefix}lti_link",
 "create table {$CFG->dbprefix}lti_link (
-    link_id             MEDIUMINT NOT NULL AUTO_INCREMENT,
+    link_id             INTEGER NOT NULL AUTO_INCREMENT,
     link_sha256         CHAR(64) NOT NULL,
     link_key            VARCHAR(4096) NOT NULL,
 
-    context_id          MEDIUMINT NOT NULL, 
+    context_id          INTEGER NOT NULL, 
 
     title               VARCHAR(2048) NULL,
 
@@ -75,12 +75,12 @@ array( "{$CFG->dbprefix}lti_link",
 
 array( "{$CFG->dbprefix}lti_user",
 "create table {$CFG->dbprefix}lti_user (
-    user_id             MEDIUMINT NOT NULL AUTO_INCREMENT,
+    user_id             INTEGER NOT NULL AUTO_INCREMENT,
     user_sha256         CHAR(64) NOT NULL,
     user_key            VARCHAR(4096) NOT NULL,
 
-    key_id              MEDIUMINT NOT NULL,
-    profile_id          MEDIUMINT NOT NULL,
+    key_id              INTEGER NOT NULL,
+    profile_id          INTEGER NOT NULL,
 
     displayname         VARCHAR(2048) NULL,
     email               VARCHAR(2048) NULL,
@@ -101,10 +101,10 @@ array( "{$CFG->dbprefix}lti_user",
 
 array( "{$CFG->dbprefix}lti_membership",
 "create table {$CFG->dbprefix}lti_membership (
-    membership_id       MEDIUMINT NOT NULL AUTO_INCREMENT,
+    membership_id       INTEGER NOT NULL AUTO_INCREMENT,
 
-    context_id          MEDIUMINT NOT NULL, 
-    user_id             MEDIUMINT NOT NULL, 
+    context_id          INTEGER NOT NULL, 
+    user_id             INTEGER NOT NULL, 
 
     role                SMALLINT NULL,
 
@@ -127,11 +127,11 @@ array( "{$CFG->dbprefix}lti_membership",
 
 array( "{$CFG->dbprefix}lti_service",
 "create table {$CFG->dbprefix}lti_service (
-    service_id          MEDIUMINT NOT NULL AUTO_INCREMENT,
+    service_id          INTEGER NOT NULL AUTO_INCREMENT,
     service_sha256      CHAR(64) NOT NULL,
     service_key         VARCHAR(4096) NOT NULL,
 
-    key_id              MEDIUMINT NOT NULL, 
+    key_id              INTEGER NOT NULL, 
 
     format              VARCHAR(1024) NULL,
 
@@ -150,14 +150,14 @@ array( "{$CFG->dbprefix}lti_service",
 
 array( "{$CFG->dbprefix}lti_result",
 "create table {$CFG->dbprefix}lti_result (
-    result_id          MEDIUMINT NOT NULL AUTO_INCREMENT,
-    link_id            MEDIUMINT NOT NULL, 
-    user_id            MEDIUMINT NOT NULL,
+    result_id          INTEGER NOT NULL AUTO_INCREMENT,
+    link_id            INTEGER NOT NULL, 
+    user_id            INTEGER NOT NULL,
 
     sourcedid          VARCHAR(2048) NOT NULL,
     sourcedid_sha256   CHAR(64) NOT NULL,
 
-    service_id         MEDIUMINT NULL,
+    service_id         INTEGER NULL,
 
     grade              FLOAT NULL,
     note               VARCHAR(2048) NOT NULL,
@@ -190,7 +190,7 @@ array( "{$CFG->dbprefix}lti_result",
 // Profile hangs out as a leaf 
 array( "{$CFG->dbprefix}profile",
 "create table {$CFG->dbprefix}profile (
-    profile_id          MEDIUMINT NOT NULL AUTO_INCREMENT,
+    profile_id          INTEGER NOT NULL AUTO_INCREMENT,
 
     displayname         VARCHAR(2048) NULL,
     email               VARCHAR(2048) NULL,
