+++ samples/blob/database.php
@@ -7,10 +7,10 @@
 $DATABASE_INSTALL = array( 
 array( "{$CFG->dbprefix}sample_blob", 
 "create table {$CFG->dbprefix}sample_blob (
-    file_id      MEDIUMINT NOT NULL KEY AUTO_INCREMENT,
+    file_id      INTEGER NOT NULL KEY AUTO_INCREMENT,
     file_sha256  CHAR(64) NOT NULL, 
 
-    context_id   MEDIUMINT NULL,
+    context_id   INTEGER NULL,
 	file_name    VARCHAR(2048),
     deleted      TINYINT(1),
     content      LONGBLOB NULL,
